@@ -8,9 +8,8 @@
 import UIKit
 import OSLog
 import Models
-import Persistence
 
-public protocol ImageCombining {
+public protocol ImageCombining: Sendable {
     func createCombinedImage(
         from images: [UIImage],
         imageQuality: Double
