@@ -13,16 +13,39 @@ struct ActionExtensionView: View {
     var body: some View {
         NavigationView {
             bodyItem
+                .task { await viewModel.loadAttachments() }
+                .alert(error: $viewModel.error) {
+                    viewModel.cancelButtonTapped()
+                }
                 .toolbar {
                     ToolbarItem(placement: .cancellationAction) {
                         Button("Cancel"){
                             viewModel.cancelButtonTapped()
-                        }
+                        }.tint(.red)
                     }
                     
-                    if let shareableImages = viewModel.shareableImages {
-                        ToolbarItem(placement: .primaryAction) {
-                            ShareLink(items: shareableImages.map(\.url))
+                    ToolbarItemGroup(placement: .primaryAction) {
+                        if viewModel.showReverseImageButton {
+                            Button {
+                                Task {
+                                    await viewModel.reverseImages()
+                                }
+                            } label: {
+                                Label("Reverse Images", systemImage: "arrow.left.arrow.right")
+                            }
+                            .disabled(viewModel.isReversingImages)
+                        }
+                        
+                        if let name = viewModel.viewTypeImageName {
+                            Button {
+                                viewModel.toggleIndividualViewType()
+                            } label: {
+                                Label("Individual View Type", systemImage: name)
+                            }
+                        }
+
+                        if let urls = viewModel.sharableURLs {
+                            ShareLink(items: urls)
                         }
                     }
                 }
@@ -32,36 +55,11 @@ struct ActionExtensionView: View {
     @ViewBuilder
     private var bodyItem: some View {
         if viewModel.canSaveFramedScreenshot {
-            mainContent
+            ActionExtensionMainContentView(viewModel: viewModel)
         } else {
             Text("Shotbot Pro Required")
                 .font(.largeTitle)
         }
     }
-    
-    private var mainContent: some View {
-        VStack {
-            if let shareableImage = viewModel.shareableImages {
-                TabView {
-                    ForEach(shareableImage) { shareableImage in
-                        Image(uiImage: shareableImage.framedScreenshot)
-                            .resizable()
-                            .scaledToFit()
-                            .padding([.horizontal, .top])
-                            .padding(.bottom, 40)
-                    }
-                }
-                .tabViewStyle(.page)
-                .onAppear {
-                    let appearance = UIPageControl.appearance()
-                    appearance.currentPageIndicatorTintColor = UIColor.gray.withAlphaComponent(0.75)
-                    appearance.pageIndicatorTintColor = UIColor.gray.withAlphaComponent(0.33)
-                }
-            } else {
-                ProgressView()
-            }
-        }
-        .navigationTitle(viewModel.title)
-        .navigationBarTitleDisplayMode(.inline)
-    }
 }
+
